Update Wsb_Scraper.py
Updated for data archival purposes. Now every day it is run a new file should be created for the day, one for the comments that have been scraped and one for the stocks mentioned on the day. Also tried to get it to work on the weekend, code doesn't run though yet.